# Conflicts:
#	app/Http/Controllers/Cooperation/Admin/Cooperation/QuestionnaireController.php

+++ app/Services/UserService.php
@@ -7,7 +7,9 @@ use App\Helpers\PicoHelper;
 use App\Models\Account;
 use App\Models\Building;
 use App\Models\BuildingFeature;
+use App\Models\CompletedQuestionnaire;
 use App\Models\Cooperation;
+use App\Models\InputSource;
 use App\Models\Role;
 use App\Models\User;
 use App\Scopes\GetValueScope;
@@ -15,6 +17,54 @@ use Illuminate\Support\Facades\Log;
 
 class UserService
 {
+
+    /**
+     * Method to reset a user his file for a specific input source
+     * 
+     * @param User $user
+     * @param InputSource $inputSource
+     */
+    public static function resetUser(User $user, InputSource $inputSource)
+    {
+        // only remove the example building id from the building
+        $building = $user->building;
+        $building->example_building_id = null;
+        $building->save();
+
+        // delete the services from a building
+        $building->buildingServices()->forInputSource($inputSource)->delete();
+        // delete the elements from a building
+        $building->buildingElements()->forInputSource($inputSource)->delete();
+        // remove the features from a building
+        $building->buildingFeatures()->forInputSource($inputSource)->delete();
+        // remove the roof types from a building
+        $building->roofTypes()->forInputSource($inputSource)->delete();
+        // remove the heater from a building
+        $building->heater()->forInputSource($inputSource)->delete();
+        // remove the solar panels from a building
+        $building->pvPanels()->forInputSource($inputSource)->delete();
+        // remove the insulated glazings from a building
+        $building->currentInsulatedGlazing()->forInputSource($inputSource)->delete();
+        // remove the paintwork from a building
+        $building->currentPaintworkStatus()->forInputSource($inputSource)->delete();
+        // remove all progress made in the tool
+        $building->completedSteps()->forInputSource($inputSource)->delete();
+        // remove the step comments
+        $building->stepComments()->forInputSource($inputSource)->delete();
+        // remove the answers on the custom questionnaires
+        $building->questionAnswers()->forInputSource($inputSource)->delete();
+
+        // remove the action plan advices from the user
+        $user->actionPlanAdvices()->forInputSource($inputSource)->delete();
+        // remove the user interests
+        $user->userInterests()->forInputSource($inputSource)->delete();
+        // remove the energy habits from a user
+        $user->energyHabit()->forInputSource($inputSource)->delete();
+        // remove the motivations from a user
+        $user->motivations()->delete();
+        // remove the progress of the completed questionnaires
+        CompletedQuestionnaire::forMe($user)->forInputSource($inputSource)->delete();
+    }
     /**
      * Method to register a user.
      *
@@ -103,34 +153,45 @@ class UserService
         return $user;
     }
 
-    public static function deleteUser(User $user)
+    /**
+     * Method to delete a user and its user info
+     *
+     * @param User $user
+     * @param bool $shouldForceDeleteBuilding
+     * @throws \Exception
+     */
+    public static function deleteUser(User $user, $shouldForceDeleteBuilding = false)
     {
         $accountId = $user->account_id;
-
         $building = $user->building;
 
-        $building->delete();
+        if ($building instanceof Building) {
+            if ($shouldForceDeleteBuilding) {
+                BuildingService::deleteBuilding($building);
+            } else {
+                $building->delete();
+                // remove the progress from a user
+                $building->completedSteps()->delete();
+            }
+        }
 
         // remove the action plan advices from the user
-        $user->actionPlanAdvices()->withoutGlobalScope(GetValueScope::class)->delete();
+        $user->actionPlanAdvices()->withoutGlobalScopes()->delete();
         // remove the user interests
-        $user->userInterests()->withoutGlobalScope(GetValueScope::class)->delete();
+        $user->userInterests()->withoutGlobalScopes()->delete();
         // remove the energy habits from a user
-        $user->energyHabit()->withoutGlobalScope(GetValueScope::class)->delete();
+        $user->energyHabit()->withoutGlobalScopes()->delete();
         // remove the motivations from a user
-        $user->motivations()->delete();
+        $user->motivations()->withoutGlobalScopes()->delete();
         // remove the notification settings
-        $user->notificationSettings()->delete();
-        // remove the progress from a user
-        $building->completedSteps()->delete();
+        $user->notificationSettings()->withoutGlobalScopes()->delete();
         // first detach the roles from the user
         $user->roles()->detach($user->roles);
-        // delete the private messages from the cooperation
-        $building->privateMessages()->delete();
 
         // remove the user itself.
         $user->delete();
 
+        // remove the user itself.
         // if the account has no users anymore then we delete the account itself too.
         if (0 == User::withoutGlobalScopes()->where('account_id', $accountId)->count()) {
             // bye !
